\name{colbycol-package}
\alias{colbycol-package}
\alias{colbycol}
\docType{package}
\title{ Reads data text files in a memory efficient way }
\description{
This package tries to solve the memory restrictions posed by large text files by breaking them into its columns, which are subsequently read individually into R.
}
\details{
\tabular{ll}{
Package: \tab colbycol\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2009-06-27\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Maintainer: Carlos J. Gil Bellosta <cgb@datanalytics.com>
}
%-\references{ }
\keyword{ package }
\keyword{ manip }
%-\seealso{ }
%-\examples{ }
