\name{colnames.colbycol}
\alias{colnames.colbycol}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Lists the column names in a colbycol object }
\description{
  \code{colnames.colbycol} lists the column names in a \code{colbycol} object.
}
\usage{
\method{colnames}{colbycol}(x, do.NULL, prefix)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a \code{colbycol} object }
  \item{do.NULL}{ required for the \code{colnames} default function }
  \item{prefix}{ required for the \code{colnames} default function }
}
%-\details{ }
\value{
    A character vector containing the names of the columns in \code{x}.
}
%-\references{ ~put references to the literature/web site here ~ }
\author{ Carlos J. Gil Bellosta }
%-\note{ }
\seealso{ \code{\link{colnames}} }
\examples{ }
\keyword{ manip }
