\name{ncol.colbycol}
\alias{ncol.colbycol}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Returns the number of cols in a colbycol object }
\description{
  This function extends the \code{\link{ncol}} function for regular \code{data.frames}.
}
\usage{
\method{ncol}{colbycol}(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ a \code{colbycol} object }
}
%-\details{ }
\value{
  An \code{integer}.
}
%-\references{ }
\author{ Carlos J. Gil Bellosta }
%-\note{ }
\seealso{ \code{\link{ncol}} }
%-\examples{ }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manip }
